Move assignment walker to resolver
